+++ package.json
@@ -100,7 +100,8 @@
     "lint:fix": "eslint . --fix",
     "test": "jest",
     "prepare": "npx husky",
-    "preinstall": "npx only-allow pnpm"
+    "preinstall": "npx only-allow pnpm",
+    "unocss": "cross-env TARO_ENV=h5 unocss \"src/**/*.{html,js,ts,jsx,tsx}\" -o src/uno.scss"
   },
   "browserslist": [
     "last 3 versions",
@@ -167,12 +168,16 @@
     "@types/webpack-env": "^1.18.8",
     "@typescript-eslint/eslint-plugin": "^8.59.1",
     "@typescript-eslint/parser": "^8.59.1",
+    "@unocss/cli": "^66.6.8",
     "@unocss/eslint-plugin": "^66.6.8",
+    "@unocss/postcss": "^66.6.8",
+    "@unocss/preset-rem-to-px": "^66.6.8",
     "@unocss/webpack": "^66.6.8",
     "babel-plugin-import": "^1.13.8",
     "babel-preset-taro": "4.2.0",
     "commitizen": "^4.3.1",
     "cross-env": "^10.1.0",
+    "css-loader": "^7.1.4",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^10.3.0",
     "eslint-config-taro": "4.2.0",
@@ -183,6 +188,7 @@
     "lint-staged": "^16.4.0",
     "postcss": "^8.5.13",
     "react-refresh": "^0.18.0",
+    "style-loader": "^4.0.0",
     "stylelint": "^17.9.1",
     "terser-webpack-plugin": "^5.5.0",
     "ts-node": "^10.9.2",
