+++ pnpm-workspace.yaml
@@ -1,0 +1,6 @@
+shellEmulator: true
+
+trustPolicy: no-downgrade
+
+patchedDependencies:
+  "@tarojs/components@4.2.0": patches/@tarojs__components@4.2.0.patch
