+++ src/import-uno.h5.ts
@@ -1,0 +1,1 @@
+import "./uno.scss";
