+++ src/import-uno.weapp.ts
@@ -1,0 +1,1 @@
+import "uno.css";
