+++ src/lib/sr/index.h5.ts
@@ -1,0 +1,7 @@
+// srConfig.ts
+export const srConfig = {};
+
+// eslint-disable-next-line unused-imports/no-unused-vars
+export async function initSRSDK(isEnabled: boolean) {
+  // H5 环境暂不加载腾讯有数小程序 SDK
+}
