+++ src/utils/updateVersion.h5.ts
@@ -1,0 +1,3 @@
+export function updateVersion() {
+  // H5 环境下不需要检查微信小程序版本更新
+}
